+++ microservice/internal/core/driven_ports/session_port.go
@@ -1,11 +1,11 @@
 package drivenPorts
 
 import (
-	"microservice/internal/domain"
+	types "microservice/internal/core/types"
 )
 
 type SessionPort interface {
 	CreateSession(userID int) (int, error)
-	GetSession(userID int) (domain.Session, error)
+	GetSession(userID int) (types.Session, error)
 	DeleteSession(sessionID int) error
 }
