+++ microservice/internal/core/driven_ports/user_port.go
@@ -1,11 +1,11 @@
 package drivenPorts
 
 import (
-	"microservice/internal/domain"
+	types "microservice/internal/core/types"
 )
 
 type UserPort interface {
-	GetUserFromID(userID int) (domain.User, error)
-	GetUserFromName(username string) (domain.User, error)
+	GetUserFromID(userID int) (types.User, error)
+	GetUserFromName(username string) (types.User, error)
 	Register(username, password, salt string) (int, error)
 }
