+++ microservice/internal/core/types/session.go
@@ -1,4 +1,4 @@
-package domain
+package types
 
 import "time"
